requirements.txt
slack-bolt
openai
python-dotenv
flask

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0